create package structure for classic
