+++ classic/__init__.py
@@ -1,0 +1,2 @@
+from .cdd import critical_difference_diagram, critical_difference_graph
+from .scatter import scatter_comparison
